requirements.txt
Pillow==9.5.0
torch==2.0.1
tqdm==4.65.0
transformers==4.28.1
Flask==2.1.1